requirements.txt
django~=3.1.11

django-extensions
django-debug-toolbar
django-bootstrap-v5
django-pydenticon
django-ckeditor
django-embed-video


pillow
easy-thumbnails

channels
channels_redis